Ray casting test

+++ lib/include/ft_math.h
@@ -6,15 +6,15 @@
 /*   By: ehosta <ehosta@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/24 13:53:02 by ehosta            #+#    #+#             */
-/*   Updated: 2025/02/13 16:20:49 by ehosta           ###   ########.fr       */
+/*   Updated: 2025/07/08 16:42:54 by ehosta           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FT_MATH_H
 # define FT_MATH_H
 
-unsigned int	ft_abs(int x);
-int				ft_sign(int x);
-int				ft_sqrt(int x);
+double	ft_abs(double x);
+int		ft_sign(int x);
+int		ft_sqrt(int x);
 
 #endif
